+++ common/SQL/CDT_COVID/Daily-stats-v2/Hospitalizations.sql
