+++ common/SQL/CDT_COVID/Daily-stats-v2/Metrics.sql
