+++ common/SQL/CDT_COVID/Daily-stats-v2/schema/tests/output/fail/deaths_null.json
