+++ common/SQL/CDT_COVID/Daily-stats-v2/schema/tests/output/fail/hospitalizations_fractional.json
