+++ common/SQL/CDT_COVID/Daily-stats-v2/schema/tests/output/fail/icu_date_format.json
