+++ common/SQL/CDT_COVID/Daily-stats-v2/schema/tests/output/fail/string_instead_of_number.json
