+++ common/SQL/CDT_COVID/Daily-stats-v2/schema/tests/output/fail/tests_new_negative.json
