+++ common/SQL/CDT_COVID/Daily-stats-v2/schema/tests/output/fail/vaccinations_0.json
