+++ common/SQL/CDT_COVID/Daily-stats-v2/schema/tests/output/pass/2021-02-22.json
@@ -1,0 +1,61 @@
+{
+  "data": {
+    "cases": {
+      "DATE": "2021-02-22",
+      "LATEST_TOTAL_CONFIRMED_CASES": 3450058,
+      "NEWLY_REPORTED_CASES": 3447,
+      "LATEST_PCT_CH_CASES_REPORTED_1_DAY": 0.1,
+      "LATEST_CONFIDENT_AVG_CASE_RATE_PER_100K_7_DAYS": 13.194893688,
+      "LATEST_CONFIDENT_INCREASE_CASE_RATE_PER_100K_7_DAYS": -9.116988744,
+      "NEWLY_REPORTED_CASES_LAST_7_DAYS": 38001
+    },
+    "deaths": {
+      "DATE": "2021-02-22",
+      "LATEST_TOTAL_CONFIRMED_DEATHS": 49129,
+      "NEWLY_REPORTED_DEATHS": 90,
+      "LATEST_CONFIDENT_AVG_DEATH_RATE_PER_100K_7_DAYS": 0.204340186,
+      "LATEST_CONFIDENT_INCREASE_DEATH_RATE_PER_100K_7_DAYS": -0.191880418,
+      "LATEST_PCT_CH_DEATHS_REPORTED_1_DAY": 0.1835
+    },
+    "tests": {
+      "DATE": "2021-02-22",
+      "LATEST_TOTAL_TESTS_PERFORMED": 46733003,
+      "LATEST_PCT_CH_TOTAL_TESTS_REPORTED_1_DAY": 0.413758,
+      "LATEST_CONFIDENT_AVG_TOTAL_TESTS_7_DAYS": 203873.571,
+      "NEWLY_REPORTED_TESTS": 192565,
+      "NEWLY_REPORTED_TESTS_LAST_7_DAYS": 1285642,
+      "LATEST_CONFIDENT_POSITIVITY_RATE_7_DAYS": 0.034266,
+      "LATEST_CONFIDENT_INCREASE_CASE_RATE_PER_100K_7_DAYS": -9.116988744,
+      "LATEST_CONFIDENT_INCREASE_DEATH_RATE_PER_100K_7_DAYS": -0.191880418,
+      "LATEST_CONFIDENT_INCREASE_POSITIVITY_RATE_7_DAYS": -0.012503
+    },
+    "hospitalizations": {
+      "DATE": "2021-02-22",
+      "HOSPITALIZED_COVID_CONFIRMED_PATIENTS": 6345,
+      "HOSPITALIZED_COVID_CONFIRMED_PATIENTS_DAILY": -224,
+      "HOSPITALIZED_COVID_CONFIRMED_PATIENTS_DAILYPCTCHG": -3.41,
+      "HOSPITALIZED_COVID_CONFIRMED_PATIENTS_LAST14DAYS": -4853,
+      "HOSPITALIZED_SUSPECTED_COVID_PATIENTS": 563,
+      "HOSPITALIZED_SUSPECTED_COVID_PATIENTS_DAILY": -33,
+      "HOSPITALIZED_SUSPECTED_COVID_PATIENTS_DAILYPCTCHG": -5.5369,
+      "HOSPITALIZED_SUSPECTED_COVID_PATIENTS_LAST14DAYS": -143
+    },
+    "icu": {
+      "DATE": "2021-02-22",
+      "ICU_COVID_CONFIRMED_PATIENTS": 1822,
+      "ICU_COVID_CONFIRMED_PATIENTS_DAILY": -111,
+      "ICU_COVID_CONFIRMED_PATIENTS_DAILYPCTCHG": -5.7424,
+      "ICU_COVID_CONFIRMED_PATIENTS_LAST14DAYS": -1342,
+      "ICU_SUSPECTED_COVID_PATIENTS": 76,
+      "ICU_SUSPECTED_COVID_PATIENTS_DAILY": -1,
+      "ICU_SUSPECTED_COVID_PATIENTS_DAILYPCTCHG": -1.2987,
+      "ICU_SUSPECTED_COVID_PATIENTS_LAST14DAYS": -24
+    },
+    "vaccinations": {
+      "DATE": "2021-02-22",
+      "DOSES_ADMINISTERED": 123456789,
+      "CUMMULATIVE_DAILY_DOSES_ADMINISTERED": 7581560,
+      "PCT_INCREASE_FROM_PRIOR_DAY": 123456789
+    }
+  }
+}
