+++ common/SQL/CDT_COVID/Daily-stats-v2/schema/tests/output/pass/2021-02-25.json
@@ -1,0 +1,61 @@
+{
+  "data": {
+    "cases": {
+      "DATE": "2021-02-25",
+      "LATEST_TOTAL_CONFIRMED_CASES": 3465726,
+      "NEWLY_REPORTED_CASES": 5400,
+      "LATEST_PCT_CH_CASES_REPORTED_1_DAY": 0.1561,
+      "LATEST_CONFIDENT_AVG_CASE_RATE_PER_100K_7_DAYS": 11.32842053,
+      "LATEST_CONFIDENT_INCREASE_CASE_RATE_PER_100K_7_DAYS": -6.830444495,
+      "NEWLY_REPORTED_CASES_LAST_7_DAYS": 37208
+    },
+    "deaths": {
+      "DATE": "2021-02-25",
+      "LATEST_TOTAL_CONFIRMED_DEATHS": 51121,
+      "NEWLY_REPORTED_DEATHS": 473,
+      "LATEST_CONFIDENT_AVG_DEATH_RATE_PER_100K_7_DAYS": 0.193303324,
+      "LATEST_CONFIDENT_INCREASE_DEATH_RATE_PER_100K_7_DAYS": -0.144889651,
+      "LATEST_PCT_CH_DEATHS_REPORTED_1_DAY": 0.9339
+    },
+    "tests": {
+      "DATE": "2021-02-25",
+      "LATEST_TOTAL_TESTS_PERFORMED": 47215212,
+      "LATEST_PCT_CH_TOTAL_TESTS_REPORTED_1_DAY": 0.385714,
+      "LATEST_CONFIDENT_AVG_TOTAL_TESTS_7_DAYS": 181135,
+      "NEWLY_REPORTED_TESTS": 181416,
+      "NEWLY_REPORTED_TESTS_LAST_7_DAYS": 1373922,
+      "LATEST_CONFIDENT_POSITIVITY_RATE_7_DAYS": 0.033126,
+      "LATEST_CONFIDENT_INCREASE_CASE_RATE_PER_100K_7_DAYS": -6.830444495,
+      "LATEST_CONFIDENT_INCREASE_DEATH_RATE_PER_100K_7_DAYS": -0.144889651,
+      "LATEST_CONFIDENT_INCREASE_POSITIVITY_RATE_7_DAYS": -0.008494
+    },
+    "hospitalizations": {
+      "DATE": "2021-02-25",
+      "HOSPITALIZED_COVID_CONFIRMED_PATIENTS": 5590,
+      "HOSPITALIZED_COVID_CONFIRMED_PATIENTS_DAILY": -344,
+      "HOSPITALIZED_COVID_CONFIRMED_PATIENTS_DAILYPCTCHG": -5.7971,
+      "HOSPITALIZED_COVID_CONFIRMED_PATIENTS_LAST14DAYS": -4282,
+      "HOSPITALIZED_SUSPECTED_COVID_PATIENTS": 562,
+      "HOSPITALIZED_SUSPECTED_COVID_PATIENTS_DAILY": -24,
+      "HOSPITALIZED_SUSPECTED_COVID_PATIENTS_DAILYPCTCHG": -4.0956,
+      "HOSPITALIZED_SUSPECTED_COVID_PATIENTS_LAST14DAYS": -60
+    },
+    "icu": {
+      "DATE": "2021-02-25",
+      "ICU_COVID_CONFIRMED_PATIENTS": 1640,
+      "ICU_COVID_CONFIRMED_PATIENTS_DAILY": -68,
+      "ICU_COVID_CONFIRMED_PATIENTS_DAILYPCTCHG": -3.9813,
+      "ICU_COVID_CONFIRMED_PATIENTS_LAST14DAYS": -1209,
+      "ICU_SUSPECTED_COVID_PATIENTS": 85,
+      "ICU_SUSPECTED_COVID_PATIENTS_DAILY": 10,
+      "ICU_SUSPECTED_COVID_PATIENTS_DAILYPCTCHG": 13.3333,
+      "ICU_SUSPECTED_COVID_PATIENTS_LAST14DAYS": 4
+    },
+    "vaccinations": {
+      "DATE": "2021-02-25",
+      "DOSES_ADMINISTERED": 123456789,
+      "CUMMULATIVE_DAILY_DOSES_ADMINISTERED": 8243711,
+      "PCT_INCREASE_FROM_PRIOR_DAY": 123456789
+    }
+  }
+}
