+++ common/SQL/CDT_COVID/Infections-by-group/infections-by-group.sql
