+++ common/SQL/CDT_COVID/Infections-by-group/schema/output/tests/fail/InvalidCategory.json
@@ -1,0 +1,12 @@
+{
+  "data": {
+    "by_age": {
+      "population": [
+        {
+          "CATEGORY": "JENNY",
+          "METRIC_VALUE": 0
+        }
+      ]
+    }
+  }
+}
