+++ common/SQL/CDT_COVID/Infections-by-group/schema/output/tests/fail/Metric-over-100.json
@@ -1,0 +1,12 @@
+{
+  "data": {
+    "by_age": {
+      "population": [
+        {
+          "CATEGORY": "0-17",
+          "METRIC_VALUE": 99999
+        }
+      ]
+    }
+  }
+}
