+++ common/SQL/CDT_COVID/Infections-by-group/schema/output/tests/fail/deaths_null.json
@@ -1,0 +1,5 @@
+{
+  "data": {
+    "deaths": null
+  }
+}
