+++ common/SQL/CDT_COVID/Infections-by-group/schema/output/tests/fail/empty.json
@@ -1,0 +1,3 @@
+{
+  "data": null
+}
