+++ common/SQL/CDT_COVID/Infections-by-group/schema/output/tests/fail/hospitalizations_fractional.json
@@ -1,0 +1,7 @@
+{
+  "data": {
+    "hospitalizations": {
+      "HOSPITALIZED_COVID_CONFIRMED_PATIENTS": 1.5
+    }
+  }
+}
