+++ common/SQL/CDT_COVID/Infections-by-group/schema/output/tests/fail/icu_date_format.json
@@ -1,0 +1,7 @@
+{
+  "data": {
+    "icu": {
+      "DATE": "2021-13-13"
+    }
+  }
+}
