+++ common/SQL/CDT_COVID/Infections-by-group/schema/output/tests/fail/string_instead_of_number.json
@@ -1,0 +1,7 @@
+{
+  "data": {
+    "cases": {
+      "NEWLY_REPORTED_CASES_LAST_7_DAYS": "bad"
+    }
+  }
+}
