+++ common/SQL/CDT_COVID/Infections-by-group/schema/output/tests/fail/tests_new_negative.json
@@ -1,0 +1,7 @@
+{
+  "data": {
+    "tests": {
+      "NEWLY_REPORTED_TESTS": -999
+    }
+  }
+}
