+++ common/SQL/CDT_COVID/Infections-by-group/schema/output/tests/fail/vaccinations_0.json
@@ -1,0 +1,7 @@
+{
+  "data": {
+    "vaccinations": {
+      "CUMMULATIVE_DAILY_DOSES_ADMINISTERED": 0
+    }
+  }
+}
