+++ common/SQL/CDT_COVID/Infections-by-group/schema/output/tests/pass/2021-02-25.json
